fix(send of struct) with int in header send basic example

+++ network/include/Header.hpp
@@ -12,7 +12,7 @@
 
 struct Header
 {
-    int8_t _id;
+    int _id;
 };
